pyproject.toml
[project]
name = "health-and-wellness-planner-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents",
    "pydantic",
    "python-dotenv",
    "streamlit",
    "openai>=1.93.0"
]
